requirements.txt
cohere
pinecone-client==5.0.1
PyPDF2==3.0.0
gradio==3.9.1
httpx==0.23.0
httpcore==0.15.0
